+++ newrelic_api/application_hosts.py
@@ -28,7 +28,8 @@ class ApplicationHosts(Resource):
         :param page: Pagination index
 
         :rtype: dict
-        :return: The JSON response of the API
+        :return: The JSON response of the API, with an additional 'pages' key
+            if there are paginated results
 
         ::
 
@@ -59,7 +60,17 @@ class ApplicationHosts(Resource):
                             "server": "integer"
                         }
                     }
-                ]
+                ],
+                "pages": {
+                    "last": {
+                        "url": "https://api.newrelic.com/v2/applications/{application_id}/hosts.json?page=2",
+                        "rel": "last"
+                    },
+                    "next": {
+                        "url": "https://api.newrelic.com/v2/applications/{application_id}/hosts.json?page=2",
+                        "rel": "next"
+                    }
+                }
             }
 
         """
@@ -148,7 +159,8 @@ class ApplicationHosts(Resource):
         :param page: Pagination index
 
         :rtype: dict
-        :return: The JSON response of the API
+        :return: The JSON response of the API, with an additional 'pages' key
+            if there are paginated results
 
         ::
 
@@ -160,7 +172,19 @@ class ApplicationHosts(Resource):
                             "string"
                         ]
                     }
-                ]
+                ],
+                "pages": {
+                    "last": {
+                        "url": "https://api.newrelic.com/v2/\
+applications/{application_id}/hosts/{host_id}/metrics.json?page=2",
+                        "rel": "last"
+                    },
+                    "next": {
+                        "url": "https://api.newrelic.com/v2/\
+applications/{application_id}/hosts/{host_id}/metrics.json?page=2",
+                        "rel": "next"
+                    }
+                }
             }
 
         """