+++ newrelic_api/application_instances.py
@@ -28,7 +28,8 @@ class ApplicationInstances(Resource):
         :param page: Pagination index
 
         :rtype: dict
-        :return: The JSON response of the API
+        :return: The JSON response of the API, with an additional 'pages' key
+            if there are paginated results
 
         ::
 
@@ -58,7 +59,17 @@ class ApplicationInstances(Resource):
                             "server": "integer"
                         }
                     }
-                ]
+                ],
+                "pages": {
+                    "last": {
+                        "url": "https://api.newrelic.com/v2/applications/{application_id}/instances.json?page=2",
+                        "rel": "last"
+                    },
+                    "next": {
+                        "url": "https://api.newrelic.com/v2/applications/{application_id}/instances.json?page=2",
+                        "rel": "next"
+                    }
+                }
             }
 
         """
@@ -158,7 +169,19 @@ class ApplicationInstances(Resource):
                             "string"
                         ]
                     }
-                ]
+                ],
+                "pages": {
+                    "last": {
+                        "url": "https://api.newrelic.com/v2/\
+applications/{application_id}/instances/{instance_id}/metrics.json?page=2",
+                        "rel": "last"
+                    },
+                    "next": {
+                        "url": "https://api.newrelic.com/v2/\
+applications/{application_id}/instances/{instance_id}/metrics.json?page=2",
+                        "rel": "next"
+                    }
+                }
             }
 
         """